+++ readstat/tests/get_var_count_test.rs
@@ -14,7 +14,7 @@ fn get_var_count() {
     let mut d = readstat::ReadStatData::new(rsp);
     let error = d.get_metadata().unwrap();
 
-    assert_eq!(error, readstat_sys::readstat_error_e_READSTAT_OK);
+    assert_eq!(error, readstat_sys::readstat_error_e_READSTAT_OK as u32);
 
     assert_eq!(d.var_count, 13);
 }
